Corrections for Linux and changes in workflows for Windows

+++ main.cpp
@@ -17,11 +17,11 @@ void insert_sorted(Container& cont, int value) {
 }
 
 template <typename Container>
-void remove_by_index(Container& cont, long unsigned int index) {
+void remove_by_index(Container& cont, size_t index) {
     if (index < 0 || index >= cont.size()) return;
 
     auto it = cont.begin();
-    for (int i = 0; i < index; ++i) {
+    for (size_t i = 0; i < index; ++i) {
         ++it;
     }
